# feature_gate: avoid extra copy when queried feature is already stored, use Set instead of map