<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->
/kind bug

#### What this PR does / why we need it:
Rework `queriedFeatures` to use `atomic.Pointer` instead of `atomic.Value` and use `sets` instead of `map[Feature]struct{}` to avoid extra allocation. `unsafeRecordQueried` is called when featuregate is queired, so this optimization ensures less memory allocations in hot-path during LIST or GET calls.

On 1.31 beta0 it allocates a lot of memory during several LISTs:
![image](https://github.com/user-attachments/assets/d4997934-dd0c-4486-ba82-f9747d8e7e86)

A simple benchmark
```
func BenchmarkUnsafeRecordQueried(b *testing.B) {
	var testFeature Feature = "Testing"
	testFeatureGate := NewFeatureGate()
	for n := 0; n < b.N; n++ {
		testFeatureGate.unsafeRecordQueried(testFeature)
	}
	b.ReportAllocs()
}
```
shows the new code is 3 times faster and does lot less allocations:
```
name                   old time/op    new time/op    delta
UnsafeRecordQueried-8     149ns ± 0%      51ns ± 0%   ~     (p=1.000 n=1+1)

name                   old alloc/op   new alloc/op   delta
UnsafeRecordQueried-8      192B ± 0%        8B ± 0%   ~     (p=1.000 n=1+1)

name                   old allocs/op  new allocs/op  delta
UnsafeRecordQueried-8      2.00 ± 0%      1.00 ± 0%   ~     (p=1.000 n=1+1)
```

With this patch allocations in the function look much better:
![image](https://github.com/user-attachments/assets/d996f7f4-0807-4618-80e1-08bc16bc60b1)


#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs
NONE
```

```release-note
NONE
```