@siyuanfoundation 2024-07-19
/lgtm Thanks for fixing this bug!